refactor: extract wp-builder and helpers from parallel-implement/workflow.ts
