+++ slices/parallel-implement/workflow.ts
@@ -11,17 +11,13 @@ import {
   shouldRunDevReview,
 } from '@goose-hub/core/agent-runtime/dev-review-advisor.js';
 import type {
-  AgentResult,
   AgentRuntime,
-  AgentSpec,
 } from '@goose-hub/core/agent-runtime/interface.js';
 import { readPromptWithContext } from '@goose-hub/core/agent-runtime/read-prompt.js';
 import { resolveGlobalSettingsForProject } from '@goose-hub/core/agent-runtime/resolve-for-project.js';
 import { toJsonSchema } from '@goose-hub/core/agent-runtime/schema-bridge.js';
 import { selectPersona } from '@goose-hub/core/agent-runtime/select-persona.js';
 import { openPR } from '@goose-hub/core/connectors/github/open-pr.js';
-import { db } from '@goose-hub/core/db/db.js';
-import { wpIterations } from '@goose-hub/core/db/schema.js';
 import type { AgentEvent, AppendEventInput } from '@goose-hub/core/event-stream/store.js';
 import { eventStore } from '@goose-hub/core/event-stream/store.js';
 import { getProjectBySlug } from '@goose-hub/core/projects/loader.js';
@@ -36,18 +32,19 @@ import {
 } from '@goose-hub/core/workspaces/orchestrator-git.js';
 import { type cleanupWorktree, createWorktree } from '@goose-hub/core/workspaces/worktree.js';
 import { ImplementWpSchema } from '@goose-hub/skills/implement-wp/schema.js';
-import type { EngineeringSpec, WorkPackage } from '@goose-hub/skills/spec-author/schema.js';
-import { and, desc, eq } from 'drizzle-orm';
+import type { EngineeringSpec } from '@goose-hub/skills/spec-author/schema.js';
+import {
+  buildParallelPrBody,
+  getLastWpStatus,
+  recordWpIteration,
+  runWithConcurrencyCap,
+  type WpDispatchResult,
+} from './parallel-helpers.js';
+import { runOneWpBuilder } from './wp-builder.js';
 
 // ─── Types ────────────────────────────────────────────────────────────────────
 
-export interface WpDispatchResult {
-  wpId: string;
-  status: 'ok' | 'failed' | 'timeout';
-  commitSha?: string;
-  errorReason?: string;
-  runId: string;
-}
+export type { WpDispatchResult } from './parallel-helpers.js';
 
 export type ParallelImplementWorkflowResult =
   | { status: 'success'; devRunId: string; worktreePath: string; prNumber: number; prUrl: string }
@@ -81,262 +78,6 @@ export interface ParallelImplementDeps {
   commitDevReviewResponseImpl?: (worktreePath: string, msg: string) => string;
   /** Override persona selection for tests that should not touch SQLite routing state. */
   selectPersonaImpl?: typeof selectPersona;
-}
-
-// ─── DB helpers ───────────────────────────────────────────────────────────────
-
-function recordWpIteration(
-  runId: string,
-  wpId: string,
-  iteration: number,
-  status: 'in-progress' | 'ok' | 'failed',
-  errorReason?: string,
-): void {
-  db.insert(wpIterations)
-    .values({ runId, wpId, iteration, status, errorReason: errorReason ?? null })
-    .onConflictDoUpdate({
-      target: [wpIterations.runId, wpIterations.wpId, wpIterations.iteration],
-      set: { status, errorReason: errorReason ?? null },
-    })
-    .run();
-}
-
-function getLastWpStatus(runId: string, wpId: string): 'ok' | 'failed' | 'in-progress' | null {
-  const rows = db
-    .select()
-    .from(wpIterations)
-    .where(and(eq(wpIterations.runId, runId), eq(wpIterations.wpId, wpId)))
-    .orderBy(desc(wpIterations.iteration))
-    .limit(1)
-    .all();
-  return (rows[0]?.status as 'ok' | 'failed' | 'in-progress' | null) ?? null;
-}
-
-// ─── Concurrency primitives ───────────────────────────────────────────────────
-
-async function runWithConcurrencyCap<T, R>(
-  items: T[],
-  cap: number,
-  fn: (item: T, index: number) => Promise<R>,
-): Promise<R[]> {
-  const results: R[] = new Array(items.length);
-  let next = 0;
-  const workerCount = Math.max(1, Math.min(cap, items.length));
-  const workers: Promise<void>[] = [];
-  for (let w = 0; w < workerCount; w++) {
-    workers.push(
-      (async (): Promise<void> => {
-        while (true) {
-          const idx = next++;
-          if (idx >= items.length) return;
-          const item = items[idx];
-          if (item === undefined) return;
-          results[idx] = await fn(item, idx);
-        }
-      })(),
-    );
-  }
-  await Promise.all(workers);
-  return results;
-}
-
-// ─── PR body ─────────────────────────────────────────────────────────────────
-
-function buildParallelPrBody(opts: {
-  workItem: WorkItem;
-  spec: EngineeringSpec;
-  wpResults: WpDispatchResult[];
-}): string {
-  const wpChangelog = opts.wpResults
-    .map((r) => {
-      const wp = opts.spec.workPackages.find((w) => w.id === r.wpId);
-      const status = r.status === 'ok' ? '✓' : '✗';
-      return `- ${status} **${r.wpId}**: ${wp?.changes.slice(0, 120) ?? '(unknown)'}`;
-    })
-    .join('\n');
-
-  return `## Summary
-
-${opts.workItem.title}
-
-## Work Packages
-
-${wpChangelog}
-
-Closes #${opts.workItem.externalId}
-`;
-}
-
-// ─── Single-WP builder runner ─────────────────────────────────────────────────
-
-// Internal result from the concurrent build phase, before the serial commit phase.
-type WpBuildPhaseResult =
-  | {
-      status: 'built';
-      wp: WorkPackage;
-      parsedFilesWritten: Array<{ path: string; reason?: string }>;
-      wpRunId: string;
-      commitMsg: string;
-      scratchWorktreePath: string;
-    }
-  | { status: 'failed' | 'timeout'; wpId: string; errorReason?: string; runId: string };
-
-interface RunOneWpBuilderOptions {
-  wp: WorkPackage;
-  iteration: number;
-  runId: string;
-  projectId: string;
-  workItemId?: string;
-  workItem: WorkItem;
-  scratchWorktreePath: string;
-  stack: { testCommand: string; lintCommand?: string; typecheckCommand?: string };
-  runtime: AgentRuntime;
-  personaId: string;
-  wpTimeoutMs: number;
-  appendEvent: (input: AppendEventInput) => AgentEvent;
-  revertWpChangesFn: typeof revertWpChanges;
-  recordIterationFn: typeof recordWpIteration;
-  implementWpPrompt: string;
-  implementWpJsonSchema: Record<string, unknown>;
-}
-
-async function runOneWpBuilder(opts: RunOneWpBuilderOptions): Promise<WpBuildPhaseResult> {
-  const { wp, iteration, runId, projectId, workItemId } = opts;
-  const wpRunId = `${runId}:wp:${wp.id}:iter:${iteration}`;
-
-  opts.recordIterationFn(runId, wp.id, iteration, 'in-progress');
-
-  opts.appendEvent({
-    projectId,
-    workItemId: workItemId ?? null,
-    kind: 'parallel-implement.wp-started',
-    payload: { wpId: wp.id, iteration, wpRunId, scratchPath: opts.scratchWorktreePath },
-    runId: wpRunId,
-  });
-
-  const spawnSpec: AgentSpec = {
-    runId: wpRunId,
-    role: 'developer',
-    skill: 'implement-wp',
-    workspaceDir: opts.scratchWorktreePath,
-    context: {
-      projectId,
-      workItemId,
-      workItem: {
-        title: opts.workItem.title,
-        body: opts.workItem.body,
-        number: Number(opts.workItem.externalId),
-        priority: opts.workItem.priority,
-      },
-      wp: {
-        id: wp.id,
-        filesOwned: wp.filesOwned,
-        changes: wp.changes,
-        dependsOn: wp.dependsOn,
-      },
-      worktreePath: opts.scratchWorktreePath,
-      stack: opts.stack,
-    },
-    contextAllowlist: [
-      'workItem.title',
-      'workItem.body',
-      'workItem.number',
-      'workItem.priority',
-      'wp.id',
-      'wp.filesOwned',
-      'wp.changes',
-      'wp.dependsOn',
-      'worktreePath',
-      'stack.testCommand',
-      'stack.lintCommand',
-      'stack.typecheckCommand',
-    ],
-    freshContext: false,
-    toolBundles: ['dev-tools'],
-    toolExtras: [],
-    budgets: { maxTurns: 150, maxBudgetUsd: 10.0, timeoutMs: opts.wpTimeoutMs },
-    personaId: opts.personaId,
-    outputJsonSchema: opts.implementWpJsonSchema,
-    appendSystemPrompt: opts.implementWpPrompt,
-    env: {
-      FACTORY_WP_FILESOWNED: wp.filesOwned.join(':'),
-      FACTORY_WP_ID: wp.id,
-    },
-  };
-
-  let result: AgentResult | undefined;
-  let errorReason: string | undefined;
-  let timedOut = false;
-
-  const start = Date.now();
-
-  try {
-    const timeout = new Promise<never>((_, reject) =>
-      setTimeout(() => reject(new Error('wp-timeout')), opts.wpTimeoutMs),
-    );
-    result = await Promise.race([opts.runtime.run(spawnSpec), timeout]);
-  } catch (err) {
-    const msg = err instanceof Error ? err.message : String(err);
-    if (msg === 'wp-timeout') {
-      timedOut = true;
-    } else {
-      errorReason = msg;
-    }
-  }
-
-  if (timedOut) {
-    opts.appendEvent({
-      projectId,
-      workItemId: workItemId ?? null,
-      kind: 'parallel-implement.wp-timeout',
-      payload: { wpId: wp.id, wpRunId, elapsedMs: Date.now() - start },
-      runId: wpRunId,
-    });
-    opts.revertWpChangesFn(opts.scratchWorktreePath, wp.filesOwned);
-    opts.recordIterationFn(runId, wp.id, iteration, 'failed', 'timeout');
-    return { status: 'timeout', wpId: wp.id, errorReason: 'timeout', runId: wpRunId };
-  }
-
-  if (errorReason != null || result == null) {
-    opts.appendEvent({
-      projectId,
-      workItemId: workItemId ?? null,
-      kind: 'parallel-implement.wp-failed',
-      payload: { wpId: wp.id, wpRunId, errorReason: errorReason ?? 'unknown' },
-      runId: wpRunId,
-    });
-    opts.revertWpChangesFn(opts.scratchWorktreePath, wp.filesOwned);
-    opts.recordIterationFn(runId, wp.id, iteration, 'failed', errorReason ?? 'unknown');
-    return { status: 'failed', wpId: wp.id, errorReason: errorReason ?? 'unknown', runId: wpRunId };
-  }
-
-  const parsed = ImplementWpSchema.safeParse(result.output);
-  if (!parsed.success) {
-    const reason = `schema validation failed: ${parsed.error.issues
-      .map((i) => `${i.path.join('.')}: ${i.message}`)
-      .join('; ')}`;
-    opts.appendEvent({
-      projectId,
-      workItemId: workItemId ?? null,
-      kind: 'parallel-implement.wp-failed',
-      payload: { wpId: wp.id, wpRunId, errorReason: reason },
-      runId: wpRunId,
-    });
-    opts.revertWpChangesFn(opts.scratchWorktreePath, wp.filesOwned);
-    opts.recordIterationFn(runId, wp.id, iteration, 'failed', reason);
-    return { status: 'failed', wpId: wp.id, errorReason: reason, runId: wpRunId };
-  }
-
-  // Build succeeded — return files for the serial commit phase.
-  const commitMsg = `M:${wp.id} ${wp.changes.slice(0, 60)}\n\nBuilt by ${opts.personaId}`;
-  return {
-    status: 'built',
-    wp,
-    parsedFilesWritten: parsed.data.filesWritten,
-    wpRunId,
-    commitMsg,
-    scratchWorktreePath: opts.scratchWorktreePath,
-  };
 }
 
 // ─── Main workflow ─────────────────────────────────────────────────────────────
